fix(mobile): move screen tests out of the expo-router app/ tree
expo-router's require.context eagerly evaluates every file under
mobile/app/ at startup. The two screen test files under app/ reference
`jest` at module top level, causing LogBox errors on device (dismissable
in dev, potential hard crash if bundled in a release build).

Relocate them to mobile/src/screens/__tests__/, outside the routed tree,
and rewrite their relative imports/mocks accordingly. jest-expo's default
testMatch still discovers them there.
